refactor: move document-outline.ts to folder
