+++ src/markdown.ts
@@ -1,13 +1,11 @@
 export {
-    type NormalizedDocumentAttributes,
     type Document,
     type DocumentBadge,
-    type FileInfo,
-} from "./document";
-export {
     type DocumentOutline,
     type DocumentOutlineEntry,
-} from "./document-outline";
+    type FileInfo,
+    type NormalizedDocumentAttributes,
+} from "./document";
 export {
     type ExampleCompiledResult,
     type ExampleStaticResult,
